refactor: update StockAPI import paths and remove unused api file
