+++ .gitignore
@@ -24,7 +24,6 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
 lib64/
 parts/
 sdist/
@@ -100,11 +99,12 @@ Thumbs.db
 *.onnx
 
 # Environment files
-*.env
+.env
 .env.local
 .env.development
 .env.test
 .env.production
+secrets.env
 
 # Project specific
 logs/
