+++ frontend/app/page.tsx
@@ -9,7 +9,7 @@ import { StockGlossary } from "@/components/stock-glossary"
 import { StockTable } from "@/components/stock-table"
 import { Top30StocksTable } from "@/components/top30-stocks-table"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { StockAPI } from "@/lib/api"
+import { StockAPI } from "@/lib/api/stock"
 import { transformRawStock } from "@/lib/utils"
 import { calculatePriceChange } from "@/lib/utils/format"
 import { useQuery } from "@tanstack/react-query"
