+++ frontend/app/stock/[tradingCode]/page.tsx
@@ -9,7 +9,7 @@ import { StockPrediction } from "@/components/stock-prediction"
 import { StockPriceChart } from "@/components/stock-price-chart"
 import { Button } from "@/components/ui/button"
 import { VolumeChart } from "@/components/volume-chart"
-import { StockAPI } from "@/lib/api"
+import { StockAPI } from "@/lib/api/stock"
 import { useQuery } from "@tanstack/react-query"
 import { ArrowLeft } from "lucide-react"
 import { useParams, useRouter } from "next/navigation"
